refactor: Update auspice_tree_to_table.py to support optional attributes
This commit updates the auspice_tree_to_table.py script in the nextstrain_data_extraction module to support optional attributes. The script now accepts a new parameter, `auspice_attributes`, which allows users to specify a list of attributes to be pulled from the auspice json file. If the parameter is specified, the script will attempt to pull only these attributes from the json file. This provides more flexibility and customization options for users when extracting data from the auspice tree.